KS-4 rework double - partOfDay

+++ src/main/kotlin/Lessons_1/task_4.kt
@@ -3,7 +3,7 @@ package org.example.Lessons_1
 fun main() {
     val length: Long = 40_868_600_000
     val age: Byte = 27
-    val partOfDay: Double = 0.075
+    val partOfDay: Float = 0.075f
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
@@ -14,7 +14,4 @@ fun main() {
     println(seconds)
     println(partOfYear)
     println(apogee)
-
-
-
 }
